+++ src/test/kotlin/io/kjson/JSONArrayTest.kt
@@ -35,8 +35,8 @@ import kotlinx.coroutines.runBlocking
 
 import java.math.BigDecimal
 
-import io.kjson.util.CoOutputCapture
-import io.kjson.util.OutputCapture
+import io.kjson.testutil.CoOutputCapture
+import io.kjson.testutil.OutputCapture
 
 class JSONArrayTest {