+++ src/test/kotlin/io/kjson/JSONBooleanTest.kt
@@ -31,8 +31,8 @@ import kotlin.test.assertTrue
 import kotlin.test.expect
 import kotlinx.coroutines.runBlocking
 
-import io.kjson.util.CoOutputCapture
-import io.kjson.util.OutputCapture
+import io.kjson.testutil.CoOutputCapture
+import io.kjson.testutil.OutputCapture
 
 class JSONBooleanTest {