+++ src/test/kotlin/io/kjson/JSONObjectTest.kt
@@ -35,8 +35,9 @@ import kotlin.test.expect
 import kotlinx.coroutines.runBlocking
 
 import java.math.BigDecimal
-import io.kjson.util.CoOutputCapture
-import io.kjson.util.OutputCapture
+
+import io.kjson.testutil.CoOutputCapture
+import io.kjson.testutil.OutputCapture
 
 class JSONObjectTest {